added finicky MS5607 driver, not working right unsure if SW/HW

+++ Firmware/T1000/digital_test/lib/ms5607/MS5607.h
@@ -4,7 +4,7 @@
 #include "stm32l4xx_ll_i2c.h"
 
 int ms5607_init(I2C_TypeDef *i2cx);
-int ms5607_get_press_temp(uint32_t double *pressure, int32_t *temperature);
+int ms5607_get_press_temp(uint32_t *pressure, int32_t *temperature);
 int ms5607_reset(void);
 
 #endif // _MS5607_H
